push

+++ DataLayer/Repositories/ContinentRepository.cs
@@ -20,18 +20,41 @@ namespace DataLayer.Repositories
 
         public void AddContinent(Continent continent)
         {
-            context.Continents.Add(continent);
+            if (!ExistsContinent(continent))
+            {
+                context.Continents.Add(continent);
+                context.SaveChanges();
+            }
+            else
+                throw new DataException("Er bestaat al een continent met deze naam!");
+            
+        }
+
+        public List<Continent> GetContinents()
+        {
+            List<Continent> continents = context.Continents.Include(c => c.Countries).ToList();
+            return continents;
         }
 
         public Continent GetContinent(int id)
         {
             Continent continent = context.Continents.Include(c =>c.Countries).FirstOrDefault(x => x.Id == id);
-            return continent;
+            if (continent != null)
+                return continent;
+            else
+                throw new DataException("No Continent for that id");
         }
 
         public void RemoveContinent(Continent continent)
         {
-            context.Continents.Remove(continent);
+            if (continent.Countries.Count == 0)
+            {
+                context.Continents.Remove(continent);
+                context.SaveChanges();
+            }
+            else
+                throw new DataException("Alle landen moeten verwijderd zijn voordat continent verwijderd kan worden");
+            
         }
 
         public void UpdateContinent(Continent continent)
@@ -48,5 +71,17 @@ namespace DataLayer.Repositories
             teUpdatenContinent.AddCountry(country);
             context.SaveChanges();
         }
+
+        public bool ExistsContinent(Continent continent)
+        {
+            bool continentFound = false;
+            foreach (Continent c in context.Continents)
+            {
+                if (c.Name.Equals(continent.Name))
+                    continentFound = true;
+            }
+            return continentFound;
+        }
+
     }
 }
